Update README.md and use argparser
Updated README to be more descriptive - lots of work yet to be done
there. Switched the argument parsing in bus2pwl so that it uses Python's
built-in argparser module instead of hacky logic. The two arguments that
argparser currently understands are the busfile required for the file to
operate and the -v (--verbose) option, which toggles whether all of the
INFO: is spewed out onto the command line when you use the script.