Add ability to expand ranges of signal values
Rewrote several parts of the script to change the way that a busfile is
parsed in order to allow ranges to be expanded. Added a couple of
example files to test the most critical cases for the range expansion.
Also added a requirement that the -p (permissive) flag be passed to the
script if invalid bus notation in the signal names is to be passed
through as a wire. If the -p flag is not passed, the script will
terminate and throw an error message now.