Improve organization of busparse/bus2pwl
Reorganized the way that the different functions in busparse call each
other in order to make the design more modular and easier to understand
and to make the different functions more available for the busverify.py
file that will be committed once it is finished. Also renamed the
'examples' folder to 'test'.